Remove pointless ascii->base64->ascii roundtrip. Commit adf97cf claimed this was for HTML sanitization but 1) it doesn't work and 2) output goes to a textarea (i.e. a string) so we don't need sanitization. Fixes #521 (unicode support) and a circular import issue.
