+++ src/util/languages/languages.js
@@ -11,7 +11,7 @@ export const SUPPORTED_LANGUAGES = [
     icon: faPython,
     codemirror: 'python',
     extension: 'py',
-    render: (code, showConsole) => CreatePythonDoc(btoa(code), showConsole),
+    render: CreatePythonDoc,
   },
   //   {
   //     value: "javascript",
@@ -65,40 +65,3 @@ export const enrichWithLanguageData = (arr) => arr.map((sketch) => ({
   ...sketch,
   language: getLanguageData(sketch.language),
 }));
-
-export const getJsSrcDocLoggingScript = () => `
-<script type="text/javascript">
-  if (typeof console  != "undefined")
-    if (typeof console.log != 'undefined')
-      console.olog = console.log;
-    else
-      console.olog = function() {};
-
-  console.log = (message) => {
-    console.olog(message);
-    let a = document.getElementById("inner")
-    if(a){
-      // a.style.display = "block"
-      a.value = a.value + "> " + message + "\\n";
-      if(a.scrollTop >= (a.scrollHeight - a.offsetHeight) - a.offsetHeight){
-        a.scrollTop = a.scrollHeight
-      }
-    }
-  };
-
-  window.onerror = (err)=> {
-    let a = document.getElementById("outer")
-    if(a){
-      a.style.display = "block"
-    }
-    console.log("\\n\\nERROR: " + err + "\\n")
-  }
-
-  console.error = console.debug = console.info = console.log;
-
-  function closeConsole(){
-    var mypre = document.getElementById("inner");
-    mypre.style.display = "none"
-  }
-</script>
-`;
